+++ app/src/main/java/com/example/findbug/domain/model/request/MemberRegisterRequestDto.kt
@@ -2,7 +2,7 @@ package com.example.findbug.domain.model.request
 
 // 고객 등록 Request
 data class MemberRegisterRequestDto(
-    val staffId: Int? = 0,
+    val staffId: Long? = 0,
     val name: String? = "",
     val email: String? = "",
     val phoneNumber: String? = "",
